refact: 맴버 변수가 아닌 직접 값을 리턴하는 구조로 변경
- 해당하는 위닝로또 관련한 테스트로 분리
